requirements.txt
# Binance API
python-binance>=1.0.19

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
pyarrow>=14.0.0

# Database
sqlalchemy>=2.0.0

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
eventlet>=0.35.0

# Async
aiohttp>=3.9.0
websockets>=12.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
